requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
defusedxml==0.7.1
distro==1.9.0
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jira==3.8.0
jiter==0.10.0
magic-filter==1.0.12
multidict==6.4.4
oauthlib==3.2.2
openai==1.82.0
packaging==25.0
pillow==11.2.1
propcache==0.3.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
sniffio==1.3.1
srt==3.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
vosk==0.3.45
websockets==15.0.1
yarl==1.20.0
pydub~=0.25.1
python-dotenv~=1.1.0
python-docx~=1.1.2